[BE] refactor: user 레포 리팩토링
- user 레포 리팩토링
