+++ server/src/domains/user/dto/UserResponse.ts
@@ -1,6 +1,6 @@
 import { IsNumber, IsString } from 'class-validator';
 
-export class UserDto {
+export class UserResponse {
   @IsNumber()
   id: number;
   @IsString()
@@ -15,17 +15,4 @@ export class UserDto {
   shareStack: number;
   @IsString()
   role?: string;
-}
-
-export class GroupUserDto {
-  @IsNumber()
-  userId: number;
-  @IsString()
-  githubId: string;
-  @IsString()
-  name: string;
-  @IsString()
-  avatarUrl: string;
-  @IsString()
-  type: string;
 }
